requirements.txt
fastapi 
uvicorn 
zxcvbn-python 
httpx